Made changes from CR to classed and gets.chomp
